+++ src/test/java/com/pnu/momeet/unit/GeneratorTest.java
@@ -1,6 +1,6 @@
 package com.pnu.momeet.unit;
 
-import com.pnu.momeet.common.security.JwtTokenProvider;
+import com.pnu.momeet.common.security.util.JwtTokenProvider;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
